Change the fontconfig to fix cache issues
The changes in commit d673b9d57eacb4ead8197fd04f3cbd99f4ec9922 broke
the generation of fontconfig caches in the runtime, because they were
written to the xdg dir instead of /usr. We still move the xdg dir
ahead of the /run/host/* dirs to ensure that at runtime the app
writes the user cache to the xdg dir.

Hopefully this will fix
 https://github.com/flatpak/flatpak/issues/1097